Buggy asf version with wrong syncronization
